delete ready just think about id when you add a service/biome

+++ app/Controllers/AdmServices.php
@@ -42,7 +42,7 @@ class AdmServices
                                 :alt
                             )
                         ');
-                        $data['upload'] = '../../' . $uploadFile;
+                        $data['upload'] = '/' . $uploadFile;
                         $query->bindValue('name', DbUtils::protectDbData($data['name']));
                         $query->bindValue('description', DbUtils::protectDbData($data['description']));
                         $query->bindValue('upload', DbUtils::protectDbData($data['upload']));
@@ -118,19 +118,31 @@ class AdmServices
         $request = json_decode($requestJSON, true);
 
         if (isset($request['id'])) {
-            $query = DbUtils::getPdo()->prepare("DELETE FROM service WHERE id = :id");
-            $query->bindValue(':id', $request['id'], \PDO::PARAM_INT);
-    
+            // get url of images
+            $query = DbUtils::getPdo()->prepare('SELECT image_url FROM service WHERE id = :id');
+            $query -> bindValue(':id', $request['id'], \PDO::PARAM_INT);
             if ($query->execute()) {
-                echo json_encode([
-                    'success' => true,
-                    'message' => 'Service supprimé avec succès.',
-                ]);
-            } else {
-                echo json_encode([
-                    'success' => false, 
-                    'message' => 'Erreur du serveur lors de l\'exécution de la requête',
-                ]);
+                $image = $query->fetch(\PDO::FETCH_ASSOC);
+            }
+            if ($image && !empty($image['image_url']) && file_exists(ltrim($image['image_url'], '/')))
+            {
+                // delete
+                $query = DbUtils::getPdo()->prepare("DELETE FROM service WHERE id = :id");
+                $query->bindValue(':id', $request['id'], \PDO::PARAM_INT);
+        
+                if ($query->execute()) {
+                    // delete images
+                    echo json_encode([
+                        'success' => true,
+                        'message' => 'Service supprimé avec succès.',
+                    ]);
+                    unlink(ltrim($image['image_url'], '/'));
+                } else {
+                    echo json_encode([
+                        'success' => false, 
+                        'message' => 'Erreur du serveur lors de l\'exécution de la requête',
+                    ]);
+                }
             }
         } else {
             echo json_encode([
